pyproject.toml
[project]
name = "mlb-leaderboards"
version = "0.1.0"
description = "Baseball Savant-style Leaderboards MVP"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "python-dotenv>=1.0.0",
    "pandas>=3.0.2",
    "yahoo-oauth>=2.1.1",
    "yahoo-fantasy-api>=2.12.3",
    "pybaseball>=2.2.7",
    "apscheduler>=3.10,<4",
    "pydantic-settings>=2.13.1",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = [
    "E501",   # line too long — docstrings and SQL are exempt
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "httpx>=0.27",
    "ruff>=0.15.10",
]
